Use hook to separate logic from UI in SignUpCard
